packages/word-processing/pyproject.toml
[build-system]
requires = ["setuptools>=80", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "vocab-tools"
dynamic = ["version"]
description = "Vocabulary processing toolkit for LinguaQuiz CEFR word lists"
readme = "README.md"
authors = [
    {name = "Nikolay Eremeev", email = "nikolay@eremeev.com"}
]
license = {text = "MIT"}
requires-python = ">=3.12"
keywords = ["vocabulary", "nlp", "language-learning", "cefr", "linguistics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Linguistic",
]

# Runtime dependencies (production)
dependencies = [
    # Core NLP dependencies
    "spacy>=3.4.0",
    "wordfreq>=3.0.0",

    # High-accuracy lemmatization (Stanford NLP)
    "stanza>=1.5.0",

    # Collocation extraction (statistical methods)
    "nltk>=3.8.0",

    # Russian morphology (specialized)
    "pymorphy3>=2.0.0",

    # Configuration management
    "PyYAML>=6.0",
    "pydantic>=2.6.0",
    "jsonschema>=4.23.0",

    # Modern CLI with Rich UI (Typer includes Rich)
    "typer[all]>=0.12.0",

    # Interactive prompts with autocomplete
    "prompt-toolkit>=3.0.43",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-xdist>=3.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

# Testing dependencies (includes property-based testing)
test = [
    "pytest>=7.0.0",
    "pytest-xdist>=3.0.0",
    "pytest-cov>=4.0.0",
    "hypothesis>=6.0.0",
    "mutmut>=2.4.0",
]

# Optional observability stack
observability = [
    "structlog>=24.0.0",
    "tenacity>=8.0.0",
]

# All optional dependencies combined
all = [
    "vocab-tools[dev,test,observability]",
]

[project.scripts]
vocab-tools = "vocab_tools.cli.app:main"

[project.urls]
Homepage = "https://github.com/nikolay-e/lingua-quiz"
Documentation = "https://github.com/nikolay-e/lingua-quiz/blob/main/packages/word-processing/CLAUDE.md"
Repository = "https://github.com/nikolay-e/lingua-quiz"
Issues = "https://github.com/nikolay-e/lingua-quiz/issues"

# Setuptools configuration
[tool.setuptools]
packages = ["vocab_tools"]

[tool.setuptools.package-data]
vocab_tools = ["py.typed", "config.yaml", "data/**/*"]

# Automatic versioning from git tags
[tool.setuptools_scm]
root = "../.."
version_scheme = "post-release"
local_scheme = "no-local-version"

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Black configuration
[tool.black]
line-length = 120
target-version = ['py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# Ruff configuration (modern linter, replaces flake8, isort, etc.)
[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = ["venv", ".venv", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by Black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex (will be addressed in refactoring)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/**/*" = ["B011"]   # assert False in tests

# MyPy configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Enable gradually during Phase 3
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

# Coverage configuration
[tool.coverage.run]
source = ["vocab_tools"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
